requirements.txt
markitdown[all]
streamlit
openai
tiktoken
markitdown
pytesseract
PyMuPDF
Pillow
